patina_debugger: Implement initial breakpoint timeout
Implements a timeout for the initial breakpoint using a timer provided by
the core. This allows for the platform to configure a timeout for the
debugger to wait for a connection before proceeding with normal execution.

This change also drops the notion that force_enable may not be used with
a timeout. This may change when a runtime enablement story is further
developed.