patina_dxe_core: Add support for reloading the core from the debugger
This commit implements a new feature that allows the Patina core to be
reloaded from the debugger without needing a full system reboot or reflash.
This is particularly useful for development and debugging, as it enables
rapid iteration on core changes. This feature is intended to be coordinated
from the UefiExt debugger extension. This feature is experimental and
is likely to evolve and change.